fix(ClearField): Moving ClearField so that it's ClearField/ClearField (#305)
