+++ react/ClearField/ClearField.js
@@ -1,7 +1,7 @@
 import styles from './ClearField.less';
 
 import React from 'react';
-import ClearIcon from '../../ClearIcon/ClearIcon';
+import ClearIcon from '../ClearIcon/ClearIcon';
 
 export default function ClearField() {
   return (
