+++ react/TextField/TextField.less
@@ -1,5 +1,5 @@
 @import (reference) "~seek-style-guide/theme";
-@import (reference) "ClearField/ClearFieldVariables";
+@import (reference) "../ClearField/ClearFieldVariables";
 
 .root {
   position: relative;
